fix: mergeability — split configuration section tests for lint line limit
Extract shared ready-state helpers and cron-specific tests so
workstation-editable-configuration-section.test.tsx stays under 450 lines.
